[modify] 불필요한 코드 삭제
- 컴포넌트로 변경
