+++ index.html
@@ -8,9 +8,9 @@
     <link rel="preload" as="style" href="/src/assets/font/pretendard/pretendard.css" />
     <link rel="stylesheet" href="/src/assets/font/pretendard/pretendard.css" />
     <link rel="stylesheet" href="/src/styles/main.css" />
+    <script type="module" src="main.js"></script>
   </head>
   <body>
     <app-root></app-root>
-    <script type="module" src="main.js"></script>
   </body>
 </html>
